requirements.txt
langchain>=0.1.2
faiss-gpu>=1.7.1
openai>=0.23.0
numpy>=1.20.0
pandas>=1.3.5
torch>=1.9.0
transformers>=4.10.0